requirements.txt
twitchio
discord.py
python-dotenv
